requirements.txt
# Core API and model integration
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
google-generativeai==0.8.3
google-auth==2.36.0
google-api-core==2.22.0

# Git integration
gitpython==3.1.41

# Evaluation metrics
evaluate==0.4.1
transformers==4.36.2
torch==2.1.2
bert-score==0.3.13
rouge-score==0.1.2
nltk==3.8.1
sentence-transformers==2.2.2

# UI
streamlit==1.30.0

# Data processing
pandas==2.1.4
numpy==1.26.3
datasets==2.16.1

# Plotting and visualization
matplotlib==3.8.2
seaborn==0.13.1

# Utilities
pyyaml==6.0.1
requests==2.31.0
tqdm==4.66.1
